refactor: move `theme-toggle` to the root of the game layout
+ renamed `(ingame)` group folder to `(gameplay)`
