+++ apps/next-app/src/app/game/(base)/layout.tsx
@@ -1,5 +1,4 @@
 import Link from "next/link"
-import dynamic from "next/dynamic"
 
 // icons
 import { LayoutDashboard } from "lucide-react"
@@ -7,21 +6,9 @@ import { LayoutDashboard } from "lucide-react"
 // shadcn
 import { Button } from "@/components/ui/button"
 
-// components
-import { ThemeToggleSkeleton } from "@/components/shared"
-
-const ThemeToggle = dynamic(() => import("@/components/shared/theme-toggle"), {
-  ssr: false,
-  loading: () => <ThemeToggleSkeleton className="bg-accent/30 rounded-full absolute top-3 right-3 sm:top-4 sm:right-4" />
-})
-
-type BaseGameLayoutProps = {
-  children: React.ReactNode
-}
-
-const BaseGameLayout = ({ children }: BaseGameLayoutProps) => {
+const BaseGameLayout = ({ children }: { children: React.ReactNode }) => {
   return (
-    <div className="relative flex-1 pt-16 pb-8 px-4 flex flex-col">
+    <>
       <Button className="bg-accent/30 hover:bg-accent/40 font-normal dark:font-light tracking-wide expandable absolute top-3 left-3 sm:top-4 sm:left-4"
         variant="ghost"
         size="icon"
@@ -35,13 +22,10 @@ const BaseGameLayout = ({ children }: BaseGameLayoutProps) => {
         </Link>
       </Button>
 
-      <ThemeToggle className="p-2 bg-accent/30 hover:bg-accent/40 absolute top-3 right-3 sm:top-4 sm:right-4"
-        variant="ghost"
-        expandable="left"
-      />
-
-      {children}
-    </div>
+      <div className="flex-1 pt-16 pb-8 px-4 flex flex-col">
+        {children}
+      </div>
+    </>
   )
 }
