+++ apps/next-app/src/app/game/(base)/summary/layout.tsx
@@ -1,20 +1,13 @@
 import Link from "next/link"
 
-// utils
-import { cn } from "@/lib/util"
-
 // icons
 import { Cog } from "lucide-react"
 
 // shadcn
-import { buttonVariants } from "@/components/ui/button"
+import { Button } from "@/components/ui/button"
 import { Separator } from "@/components/ui/separator"
 
-type BaseGameSummaryLayoutProps = {
-  children: React.ReactNode
-}
-
-const BaseGameSummaryLayout = ({ children }: BaseGameSummaryLayoutProps) => {
+const BaseGameSummaryLayout = ({ children }: { children: React.ReactNode }) => {
   return (
     <>
       <header className="mt-4">
@@ -28,16 +21,21 @@ const BaseGameSummaryLayout = ({ children }: BaseGameSummaryLayoutProps) => {
           Your <span className="text-accent font-medium">game is over</span>, thanks for playing!
         </p>
 
-        <Link className={cn(buttonVariants({
-          className: "w-fit mt-2 mx-auto flex gap-x-2 border-border/40 hover font-heading tracking-wider hover:bg-accent/15 dark:hover:bg-accent/10 hover:text-foreground",
-          variant: "outline",
-          size: "sm"
-        }))}
-          href="/game/setup"
+        <Button className="w-fit mt-2 mx-auto flex gap-x-2 border-border/40 hover font-heading tracking-wider hover:bg-accent/15 dark:hover:bg-accent/10 hover:text-foreground"
+          variant="outline"
+          size="sm"
+          asChild
         >
-          <Cog className="size-4 sm:size-5" strokeWidth={1.75} />
-          <span className="mt-1">New game</span>
-        </Link>
+          <Link href="/game/setup">
+            <Cog className="size-4 sm:size-5 shrink-0"
+              strokeWidth={1.75}
+            />
+
+            <span className="mt-1">
+              New game
+            </span>
+          </Link>
+        </Button>
       </header>
 
       <Separator className="w-3/5 mx-auto mt-4 mb-6 bg-border/5" />
