+++ apps/next-app/src/app/game/(gameplay)/layout.tsx
@@ -1,0 +1,9 @@
+const GameplayLayout = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <div className="flex-1 flex flex-col">
+      {children}
+    </div>
+  )
+}
+
+export default GameplayLayout
