+++ apps/next-app/src/app/game/(gameplay)/multiplayer/layout.tsx
