+++ apps/next-app/src/app/game/(gameplay)/offline/offline-game-handler.tsx
