+++ apps/next-app/src/app/game/(gameplay)/offline/page.tsx
