+++ apps/next-app/src/app/game/(gameplay)/single/page.tsx
