+++ apps/next-app/src/app/game/(gameplay)/single/single-game-handler.tsx
