--- apps/next-app/src/app/game/(in-game)/layout.tsx
@@ -1,13 +1,0 @@
-type InGameLayoutProps = {
-  children: React.ReactNode
-}
-
-const InGameLayout = ({ children }: InGameLayoutProps) => {
-  return (
-    <div className="flex-1 flex flex-col">
-      {children}
-    </div>
-  )
-}
-
-export default InGameLayout
